requirements.txt
numpy==1.26.2
opencv_python==4.8.1.78
Pillow==10.2.0
protobuf==4.25.1
python-dotenv==1.0.0
streamlit==1.29.0
streamlit_webrtc==0.47.1